fix: don't skip history items when running full history pull after session-specific pull (#5826)
* fix: don't skip history items when running full history pull after session-specific pull

Previously, only a single timestamp was tracked for the most recent history pull. Any pull would reset
the timestamp, regardless of which session it was pulling from. This meant that if you ran a pull with
a specific session id, then ran a general pull with no session id, you would only get history items
from after the session-specific pull, even ones in other sessions (that you didn't pull from in the first
place).

Now we track a timestamp for each individually-pulled session id, plus a timestamp for the last general
pull.

In theory there's no limit on how large the mapping of session ids to timestmaps can get, but in practice
it's unlikely to be a problem since you will typically only be pulling from a few particular ids
(i.e. whatever other sessions you currently have open) in a given session.

* fix test_history_pull_mixed on windows

* even more fixes for test_history_pull_mixed on Windows

---------

Co-authored-by: Andy Kipp <anki-code@users.noreply.github.com>